update: full pipeline
