+++ src/consts.py
@@ -1,9 +1,12 @@
 from pathlib import Path
 
-ROOT = Path(__file__).parent.parent
+ROOT = Path(__file__).parent.parent.parent
+WORKSPACE = Path(__file__).parent.parent
 DIR_TRANS = ROOT / "trans"
 DIR_SOURCE = ROOT / "source"
+DIR_SOURCE = ROOT / "BDCC"
 DIR_FETCH = ROOT / "fetch"
 DIR_OUTPUT = ROOT / "output"
+DIR_DEPRECATED = ROOT / "deprecated"
